renamed contiter, moved formating
