+++ pol-core/bscript/bdict.cpp
@@ -15,11 +15,11 @@
 
 #include "../clib/stlutil.h"
 #include "barray.h"
+#include "bcontiter.h"
 #include "berror.h"
 #include "blong.h"
 #include "bstring.h"
 #include "buninit.h"
-#include "contiter.h"
 #include "executor.h"
 #include "objmethods.h"
