+++ pol-core/bscript/bstring.cpp
@@ -23,6 +23,7 @@
 #include "../clib/strutil.h"
 #include "barray.h"
 #include "bboolean.h"
+#include "bcontiter.h"
 #include "bdouble.h"
 #include "berror.h"
 #include "bfuncref.h"
@@ -30,7 +31,6 @@
 #include "bobject.h"
 #include "bregexp.h"
 #include "buninit.h"
-#include "contiter.h"
 #include "executor.h"
 #include "objmethods.h"
 #include "str.h"
