+++ pol-core/bscript/bstruct.cpp
@@ -17,12 +17,12 @@
 #include "../clib/passert.h"
 #include "../clib/stlutil.h"
 #include "barray.h"
+#include "bcontiter.h"
 #include "berror.h"
 #include "blong.h"
 #include "bobject.h"
 #include "bstring.h"
 #include "buninit.h"
-#include "contiter.h"
 #include "executor.h"
 #include "objmethods.h"
