+++ pol-core/pol/sqlscrobj.cpp
@@ -10,12 +10,12 @@
 #include <regex>
 #include <string.h>
 
+#include "bscript/bcontiter.h"
 #include "bscript/bdouble.h"
 #include "bscript/berror.h"
 #include "bscript/blong.h"
 #include "bscript/bstring.h"
 #include "bscript/buninit.h"
-#include "bscript/contiter.h"
 #include "bscript/objmembers.h"
 #include "bscript/objmethods.h"
 #include "clib/esignal.h"
